[skip ci] reformat desktop dashboard
