+++ cw_monero/example/.gitignore
@@ -30,6 +30,7 @@ migrate_working_dir/
 .packages
 .pub-cache/
 .pub/
+/pubspec.lock
 /build/
 
 # Symbolication related
